Move partial characters to src
