--- experimental/Rings/src/Rings.jl
@@ -1,1 +1,0 @@
-include("QQAbAndPChars.jl")
