--- experimental/Rings/test/runtests.jl
@@ -1,3 +1,0 @@
-@testset "dummy test" begin
-  @test 1 + 1 == 2
-end
