+++ src/Rings/binomial_ideals.jl
@@ -359,13 +359,136 @@ end
 #
 ###################################################################################
 
-function ideal_from_character(P::QQAbModule.PartialCharacter, R::MPolyRing)
+mutable struct PartialCharacter{T}
+  #A has generators of the lattice in rows
+  A::ZZMatrix
+  #images of the generators are saved in b
+  b::Vector{T}
+  #Delta are the indices of the cellular variables of the associated ideal
+  #(the partial character is a partial character on Z^Delta)
+  D::Set{Int64}
+  function PartialCharacter{T}() where T 
+    return new{T}()
+  end
+
+  function PartialCharacter{T}(mat::ZZMatrix, vals::Vector{T}) where T
+    z = new{T}()
+    z.A = mat
+    z.b = vals
+    return z
+  end
+end
+
+function partial_character(A::ZZMatrix, vals::Vector{T}, variables::Set{Int} = Set{Int}()) where T <: FieldElem
+  @assert nrows(A) == length(vals)
+  z = PartialCharacter{T}(A, vals)
+  if !isempty(variables)
+    z.D = variables
+  end
+  return z
+end
+
+function (Chi::PartialCharacter)(b::ZZMatrix)
+  @assert nrows(b) == 1
+  @assert Nemo.ncols(b) == Nemo.ncols(Chi.A)
+  s = solve(Chi.A, b, side = :left)
+  return evaluate(FacElem(Dict([(Chi.b[i], s[1, i]) for i = 1:length(Chi.b)])))
+end
+
+function (Chi::PartialCharacter)(b::Vector{ZZRingElem})
+  return Chi(matrix(FlintZZ, 1, length(b), b))
+end
+
+function have_same_domain(P::PartialCharacter, Q::PartialCharacter)
+  return have_same_span(P.A, Q.A)
+end
+
+function have_same_span(A::ZZMatrix, B::ZZMatrix)
+  @assert ncols(A) == ncols(B)
+  return hnf(A) == hnf(B)
+end
+
+
+
+function Base.:(==)(P::PartialCharacter{T}, Q::PartialCharacter{T}) where T <: FieldElem
+  if P === Q
+    return true
+  end
+  if !have_same_domain(P, Q)
+    return false
+  end
+  #now test if the values taken on the generators of the lattices are equal
+  for i = 1:nrows(P.A)
+    TestVec = view(P.A, i:i, 1:Nemo.ncols(P.A))
+    if P(TestVec) != Q(TestVec)
+      return false
+    end
+  end
+  return true
+end
+
+function saturations(L::PartialCharacter{QQAbElem{T}}) where T
+  #computes all saturations of the partial character L
+  res = PartialCharacter{QQAbElem{T}}[]
+
+  #first handle case where the domain of the partial character is the zero lattice
+  #in this case return L
+  if iszero(L.A)
+    push!(res, L)
+    return res
+  end
+
+  #now not trivial case
+  H = hnf(transpose(L.A))
+  H = view(H, 1:ncols(H), 1:ncols(H))
+  i, d = pseudo_inv(H)  #iH = d I_n
+  #so, saturation is i' * H // d
+  S = divexact(transpose(i)*L.A, d)
+
+  B = Vector{Vector{QQAbElem{T}}}()
+  for k = 1:nrows(H)
+    c = i[1, k]
+    for j = 2:ncols(H)
+      c = gcd(c, i[j, k])
+      if isone(c)
+        break
+      end
+    end
+    mu = evaluate(FacElem(Dict(Tuple{QQAbElem{T}, ZZRingElem}[(L.b[j], div(i[j, k], c)) for j = 1:ncols(H)])))
+    mu1 = roots(mu, Int(div(d, c)))
+    push!(B,  mu1)
+  end
+  it = Hecke.cartesian_product_iterator(UnitRange{Int}[1:length(x) for x in B])
+  vT = Vector{Vector{QQAbElem{T}}}()
+  for I in it
+    push!(vT, [B[i][I[i]] for i = 1:length(B)])
+  end
+  
+  for k = 1:length(vT)
+    #check if PChar(S,vT[k],L.D) puts on the right value on the lattice generators of L
+    Pnew = partial_character(S, vT[k], L.D)
+    flag = true   #flag if value on lattice generators is right
+    for i = 1:Nemo.nrows(L.A)
+      if Pnew(sub(L.A, i:i ,1:Nemo.ncols(L.A))) != L.b[i]
+        flag = false
+        println("found wrong saturation (for information), we delete it")
+      end
+    end
+    if flag
+      push!(res, Pnew)
+    end
+  end
+  return res
+end
+
+
+function ideal_from_character(P::PartialCharacter, R::MPolyRing)
   #input: partial character P and a polynomial ring R
   #output: the ideal $I_+(P)=\langle x^{u_+}- P(u)x^{u_-} \mid u \in P.A \rangle$
 
   @assert ncols(P.A) == nvars(R)
   #test if the domain of the partial character is the zero lattice
-  if isone(nrows(P.A)) && QQAbModule.have_same_span(P.A, zero_matrix(FlintZZ, 1, ncols(P.A)))
+  if isone(nrows(P.A)) && have_same_span(P.A, zero_matrix(FlintZZ, 1, ncols(P.A)))
     return ideal(R, zero(R))
   end
 
@@ -425,7 +548,7 @@ function ideal_from_character(P::QQAbModule.PartialCharacter, R::MPolyRing)
   return I;
 end
 
-function _make_binomials(P::QQAbModule.PartialCharacter, R::MPolyRing)
+function _make_binomials(P::PartialCharacter, R::MPolyRing)
   #output: ideal generated by the binomials corresponding to the generators of the domain P.A of the partial character P
   #Note: This is not the ideal I_+(P)!!
   @assert ncols(P.A) == nvars(R)
@@ -465,7 +588,7 @@ function partial_character_from_ideal(I::MPolyIdeal, R::MPolyRing)
 
   Delta = cell[2]   #cell variables
   if isempty(Delta)
-    return QQAbModule.partial_character(zero_matrix(FlintZZ, 1, nvars(R)), [one(QQAb)], Set{Int64}())
+    return partial_character(zero_matrix(FlintZZ, 1, nvars(R)), [one(QQAb)], Set{Int64}())
   end
 
   #now consider the case where Delta is not empty
@@ -481,7 +604,7 @@ function partial_character_from_ideal(I::MPolyIdeal, R::MPolyRing)
   end
   QQAbcl, = abelian_closure(QQ)
   if iszero(J)
-    return QQAbModule.partial_character(zero_matrix(FlintZZ, 1, nvars(R)), [one(QQAbcl)], Set{Int64}())
+    return partial_character(zero_matrix(FlintZZ, 1, nvars(R)), [one(QQAbcl)], Set{Int64}())
   end
   #now case if J \neq 0
   #let ts be a list of minimal binomial generators for J
@@ -510,7 +633,7 @@ function partial_character_from_ideal(I::MPolyIdeal, R::MPolyRing)
     i -= 1
   end
   vs = view(vs, 1:i, 1:nvars(R))
-  return QQAbModule.partial_character(vs, images, Set{Int64}(Delta))
+  return partial_character(vs, images, Set{Int64}(Delta))
 end
 
 ###################################################################################
@@ -727,7 +850,7 @@ function cellular_associated_primes(I::MPolyIdeal{QQMPolyRingElem}, RQQAb::MPoly
     Im = quotient(I, ideal(R, m))
     Pm = partial_character_from_ideal(Im, R)
     #now compute all saturations of the partial character Pm
-    PmSat = QQAbModule.saturations(Pm)
+    PmSat = saturations(Pm)
     for P in PmSat
       new_id = ideal_from_character(P, RQQAb) + idealDeltaC
       push!(associated_primes, new_id)
@@ -795,7 +918,7 @@ function cellular_minimal_associated_primes(I::MPolyIdeal{QQMPolyRingElem})
   P = partial_character_from_ideal(I, R)
   QQAbcl, = abelian_closure(QQ)
   RQQAb = polynomial_ring(QQAbcl, symbols(R))[1]
-  PSat = QQAbModule.saturations(P)
+  PSat = saturations(P)
   minimal_associated = Vector{MPolyIdeal{Generic.MPoly{QQAbElem{AbsSimpleNumFieldElem}}}}() #this will hold the set of minimal associated primes
 
   #construct the ideal (x_i \mid i \in \Delta^c)
